requirements.txt
numpy
numba
pandas
scipy
pyyaml
yfinance
# technical analysis, indicator libraries
ta
# ta-lib
# bta-lib
pandas-ta
# tulipy
# broker API libraries
tda-api
cryptography
# visualization
plotly
cufflinks
nbformat
# Webdriver
chromedriver-binary

